// src/shared/enums/SocialLinks.ts
export enum SocialLinks {
    Facebook = "https://www.facebook.com/emoney.ge/",
    YouTube = "https://www.youtube.com/channel/UCkCpwAAgq7wR29jTDG3bXgw",
    Tweeter = "https://x.com/eMoneyge?mx=2",
    Instagram = "https://www.instagram.com/emoney.georgia/",
    Telegram = "https://telegram.me/emoneygeorgia",
    LinkedIn = "https://www.linkedin.com/company/emoney-ge/",
    App = "https://getapp.emoney.ge/download"
}